// apps/frontend/src/api/generated/models/videoResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Route Streaming API.
 * Route Streaming Service Backend API.
 * OpenAPI spec version: 1.0.
 */

export interface VideoResponse {
  description?: string
  duration_ms?: number
  file_path?: string
  file_size?: number
  id?: string
  mime_type?: string
  owner_id?: string
  status?: string
  title?: string
  visibility?: string
}